requirements.txt
aiosqlite==0.21.0
certifi==2025.8.3
cffi==2.0.0
curl_cffi==0.13.0
greenlet==3.2.4
numpy==2.3.3
pandas==2.3.2
pillow==11.3.0
playwright==1.55.0
pycparser==2.23
pyee==13.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
xlsxwriter==3.2.9
